Separate test from suite setup/teardown
.. in preparation for having more tests, and more setup and teardown
(specifically: setting a local container registry up).
